+++ internal/domain/repository/grape_type_repository.go
@@ -8,5 +8,5 @@ import (
 
 type GrapeTypeRepository interface {
 	AddGrapeType(ctx context.Context, grapeType *entity.GrapeType) error
-	GetGrapeType(ctx context.Context, grapeTypeId int) (*entity.GrapeType, error)
+	GetGrapeType(ctx context.Context, grapeTypeId int64) (*entity.GrapeType, error)
 }
