+++ internal/domain/repository/process_repository.go
@@ -7,7 +7,7 @@ import (
 )
 
 type ProcessRepository interface {
-	GetProcess(ctx context.Context, id uint) (*entity.Process, error)
+	GetProcess(ctx context.Context, id int64) (*entity.Process, error)
 	GetAllProcesses() ([]*entity.Process, error)
 	AddProcess(ctx context.Context, process *entity.Process) (*entity.Process, error)
 }
