+++ internal/domain/repository/reception_repository.go
@@ -8,7 +8,7 @@ import (
 
 type ReceptionRepository interface {
 	AddReception(ctx context.Context, rc *entity.Reception) (*entity.Reception, error)
-	GetReception(ctx context.Context, receptionId int) (*entity.Reception, error)
+	GetReception(ctx context.Context, receptionId int64) (*entity.Reception, error)
 	GetAllReceptions() ([]*entity.Reception, error)
-	UpdateReception(ctx context.Context, receptionId int, updateFn func(ctx context.Context, rc *entity.Reception) (*entity.Reception, error)) error
+	UpdateReception(ctx context.Context, receptionId int64, updateFn func(ctx context.Context, rc *entity.Reception) (*entity.Reception, error)) error
 }
