+++ internal/domain/repository/truck_repository.go
@@ -8,5 +8,5 @@ import (
 
 type TruckRepository interface {
 	AddTruck(ctx context.Context, truck *entity.Truck) (*entity.Truck, error)
-	GetTruck(ctx context.Context, truckId int) (*entity.Truck, error)
+	GetTruck(ctx context.Context, truckId int64) (*entity.Truck, error)
 }
